Refactor: move ObjectPermission to separate module
