+++ mwdb/model/__init__.py
@@ -51,7 +51,8 @@ from .file import File  # noqa: E402
 from .group import Group, Member  # noqa: E402
 from .karton import KartonAnalysis, karton_object  # noqa: E402
 from .oauth import OpenIDProvider, OpenIDUserIdentity  # noqa: E402
-from .object import Object, ObjectPermission, relation  # noqa: E402
+from .object import Object, relation  # noqa: E402
+from .object_permission import ObjectPermission  # noqa: E402
 from .quick_query import QuickQuery  # noqa: E402
 from .tag import Tag  # noqa: E402
 from .user import User  # noqa: E402
